Refactoring of multiple field operators stencils (#47)
Re-factoring of multiple field operators:

* mo_nh_diffusion_stencil_16, mo_nh_diffusion_stencil_08, mo_intp_rbf_rbf_vec_interpol_vertex_dom_b

* mo_solve_nonhydro_stencil_07, mo_solve_nonhydro_stencil_13

* mo_solve_nonhydro_stencil_14

* mo_solve_nonhydro_stencil_32, mo_solve_nonhydro_stencil_33

* mo_solve_nonhydro_stencil_41, mo_solve_nonhydro_stencil_42

* mo_solve_nonhydro_stencil_47, mo_solve_nonhydro_stencil_50

+++ atm_dyn_iconam/tests/test_mo_nh_diffusion_stencil_06.py
@@ -32,7 +32,7 @@ def mo_nh_diffusion_stencil_06_numpy(
 def test_mo_nh_diffusion_stencil_06():
     mesh = SimpleMesh()
 
-    fac_bdydiff_v = np.float64(5.0)
+    fac_bdydiff_v = 5.0
     z_nabla2_e = random_field(mesh, EdgeDim, KDim)
     area_edge = random_field(mesh, EdgeDim)
     vn = random_field(mesh, EdgeDim, KDim)
